cloudrun-service2service/gateway-service/requirements.txt
Flask==3.0.0
requests==2.31.0
google-auth==2.25.2
gunicorn==21.2.0